feat(ovpack): add vector snapshots and consistency checks
