+++ openviking/storage/ovpack/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2026 Beijing Volcano Engine Technology Co., Ltd.
+# SPDX-License-Identifier: AGPL-3.0
+"""OVPack internal helpers."""
